+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "@types/react-transition-group": "^4.4.4",
     "boostrap": "^2.0.0",
     "react": "^17.0.2",
     "react-addons-css-transition-group": "^15.6.2",
