+++ src/components/Pages/Main/Main.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import CardButton from "../../CardButton/CardButton";
 import {Link} from "react-router-dom";
-import {mainMenu} from "../../../State/State";
+import {mainMenu} from "../../../content/MainMenu";
 import s from './Main.module.css';
 import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
 
@@ -9,9 +9,9 @@ const Main = (props) => {
     document.title = props.title
     document.querySelector('meta[name="description"]').content = props.description
     document.documentElement.scrollIntoView(true);
-    const Cardbuttons = mainMenu.items.map( (item) => {
+    let Cardbuttons = mainMenu.items.map( (item, index) => {
         return (
-            <div className={s.card + " col-4 col-md-3 col-xl-2"}><Link to={item.urm}><CardButton name={item.name} img={item.img}/></Link></div>
+            <div key={index} className={s.card + " col-4 col-md-3 col-xl-2"}><Link to={item.urm}><CardButton name={item.name} img={item.img}/></Link></div>
         )
     })
