+++ src/components/Pages/TelegramBot/TelegramBot.js
@@ -14,7 +14,6 @@ const TelegramBot = (props) => {
                 transitionAppearTimeout={1000}
                 transitionEnter={false}
                 transitionLeave={false}>
-            {/*<hr/>*/}
                 <Titleh1 name={'Разработка приложений (ботов) для Telegram'} subname={'любой сложности'} />
                 <img src="/images/bot-telegram.png" width={'100%'} alt="DICOM разработка ботов любой сложности"/>
             </ReactCSSTransitionGroup>
