+++ src/content/MainMenu.js
@@ -1,0 +1,14 @@
+export let mainMenu = {
+    items: [
+        {name: 'Услуги', urm: '/uslugi', img: '/images/dicomYslugi.png'},
+        {name: 'Кейсы', urm: '/cases', img: '/images/dicomCase.png'},
+        {name: 'Команда', urm: '/team', img: '/images/dicomCommand.png'},
+        {name: 'Контакты, реквизиты', urm: '/contacts', img: '/images/dicomContacts.png'},
+        {name: 'Вакансии', urm: '/vacancys', img: '/images/dicomVakan.png'},
+        // {name: 'Support', urm: '/promotion', img: '/images/dicomHelp.png'},
+        {name: 'Битрикс24', urm: '/bitrix24', img: '/images/dicomB24.png'},
+        {name: '1C', urm: '/1c', img: '/images/dicom1c.png'},
+        {name: 'Telegram bots', urm: '/telegram-bot', img: '/images/telegram.png'},
+        {name: 'Сервисы', urm: '/services', img: '/images/icon1.png'}
+    ]
+}
