+++ src/content/Team.js
@@ -1,0 +1,9 @@
+export const TeamList = {
+    items: [
+        {name: 'Алексей', skills: 'Системный архитектор', photo: '/images/team/Alex.png'},
+        {name: "Вадим", skills: 'Team leader, Backend-developer, Frontend developer', photo: '/images/team/Vadim.jpg'},
+        {name: 'Сергей', skills: 'Python Backend-developer, telegram bot, QA-тестировщик', photo: '/images/team/python.png'},
+        {name: 'Аркадий', skills: 'JS Backend-developer Express, JS Frontend developer React', photo: '/images/team/js.png'},
+        {name: 'ДАРЬЯ', skills: 'UI/UX дизайнер', photo: ''}
+    ]
+}
